bake: move userfunc
