+++ bake/hclparser/hclparser.go
@@ -11,8 +11,8 @@ import (
 	"strconv"
 	"strings"
 
+	"github.com/docker/buildx/bake/hclparser/hcl/ext/userfunc"
 	"github.com/docker/buildx/bake/hclparser/hcl/gohcl"
-	"github.com/docker/buildx/util/userfunc"
 	"github.com/hashicorp/hcl/v2"
 	"github.com/pkg/errors"
 	"github.com/zclconf/go-cty/cty"
